fix crash when loading certain emotes
this was a problem that plagued this client for quite a while.
i was writing a streaming-decompression solution but that wasn't it.
turns out i was allocating way too many bytes on charData than what was needed, by using the size of the frameIndexes vector, instead of how many non-duplicate frames are in the .img.bin file...